requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
gradio
sounddevice
numpy
webrtcvad
scipy
gtts
playsound
torch
transformers
faster-whisper
accelerate
